// src/code/api.ts
export enum ProfileType {
  UCS = 0,
  ED,
  LC,
  SnN
}

export enum FogOfWarType {
  Black = 0,
  Environmental,
}

export enum PanelPosition {
  Top = 0,
  Bottom,
  Left,
  Right,
}

export enum ToolbarPosition {
  Top = 0,
  Bottom,
}

export enum MapPosition {
  TopLeft = 0,
  BottomRight,
}

export enum SelectionShape {
  Square = 0,
  Isometric,
}

export enum DescriptionType {
  Normal = 0,
  Tooltip,
}
export enum ScreenInfo {
  None = 0,
  Clock,
  Temperature,
}

export enum ShieldType {
  None = -1,
  Small,
  Medium,
  Large,
}

export enum DIK {
  Escape = 0x01,
  D1 = 0x02,
  D2 = 0x03,
  D3 = 0x04,
  D4 = 0x05,
  D5 = 0x06,
  D6 = 0x07,
  D7 = 0x08,
  D8 = 0x09,
  D9 = 0x0A,
  D0 = 0x0B,
  Minus = 0x0C,
  Equals = 0x0D,
  Backspace = 0x0E,
  Tab = 0x0F,
  Q = 0x10,
  W = 0x11,
  E = 0x12,
  R = 0x13,
  T = 0x14,
  Y = 0x15,
  U = 0x16,
  I = 0x17,
  O = 0x18,
  P = 0x19,
  LeftBracket = 0x1A,
  RightBracket = 0x1B,
  Enter = 0x1C,
  LeftControl = 0x1D,
  A = 0x1E,
  S = 0x1F,
  D = 0x20,
  F = 0x21,
  G = 0x22,
  H = 0x23,
  J = 0x24,
  K = 0x25,
  L = 0x26,
  Semicolon = 0x27,
  Apostrophe = 0x28,
  Grave = 0x29,
  LeftShift = 0x2A,
  Backslash = 0x2B,
  Z = 0x2C,
  X = 0x2D,
  C = 0x2E,
  V = 0x2F,
  B = 0x30,
  N = 0x31,
  M = 0x32,
  Comma = 0x33,
  Period = 0x34,
  Slash = 0x35,
  RightShift = 0x36,
  NumpadMultiply = 0x37,
  LeftAlt = 0x38,
  Space = 0x39,
  CapsLock = 0x3A,
  F1 = 0x3B,
  F2 = 0x3C,
  F3 = 0x3D,
  F4 = 0x3E,
  F5 = 0x3F,
  F6 = 0x40,
  F7 = 0x41,
  F8 = 0x42,
  F9 = 0x43,
  F10 = 0x44,
  NumLock = 0x45,
  ScrollLock = 0x46,
  Numpad7 = 0x47,
  Numpad8 = 0x48,
  Numpad9 = 0x49,
  NumpadMinus = 0x4A,
  Numpad4 = 0x4B,
  Numpad5 = 0x4C,
  Numpad6 = 0x4D,
  NumpadPlus = 0x4E,
  Numpad1 = 0x4F,
  Numpad2 = 0x50,
  Numpad3 = 0x51,
  Numpad0 = 0x52,
  NumpadPeriod = 0x53,
  F11 = 0x57,
  F12 = 0x58,
  F13 = 0x64,
  F14 = 0x65,
  F15 = 0x66,
  Kana = 0x70,
  Convert = 0x79,
  NoConvert = 0x7B,
  Yen = 0x7D,
  NumpadEquals = 0x8D,
  Circumflex = 0x90,
  At = 0x91,
  Colon = 0x92,
  Underline = 0x93,
  Kanji = 0x94,
  Stop = 0x95,
  AX = 0x96,
  Unlabeled = 0x97,
  NumpadEnter = 0x9C,
  RightControl = 0x9D,
  NumpadDivide = 0xB5,
  SysRq = 0xB7,
  RightAlt = 0xB8,
  Pause = 0xC5,
  Home = 0xC7,
  UpArrow = 0xC8,
  PageUp = 0xC9,
  LeftArrow = 0xCB,
  RightArrow = 0xCD,
  End = 0xCF,
  DownArrow = 0xD0,
  PageDown = 0xD1,
  Insert = 0xD2,
  Delete = 0xD3,
  LeftWindows = 0xDB,
  RightWindows = 0xDC,
  Menu = 0xDD,
  Power = 0xDE,
  Sleep = 0xDF,
}

export interface TemplateData {
  field_0x4: number,
  chassisName: string,
  field_0xc: number,
  field_0x10: number,
  field_0x14: number,
  powerShieldType: ShieldType,
  scriptId: string,
  slots: string[],
  templateName: string,
}

export interface ShortcutData {
  showInTooltip: boolean;
  boundKey: DIK;
  modifierKey: DIK;
}

export interface GameOptions {
  fogOfWar: FogOfWarType,
  atmosphericFog: boolean,
  tunnelWalls: boolean,
  tunnelWallsTransparent: boolean,
  objectShadows: boolean,
  terrainShadows: boolean,
  rain: boolean,
  snow: boolean,
  lightCons: boolean,
  lowResTextures: boolean,
  shining: boolean,
  gamma1: number,
  gamma2: number,
  gamma3: number,
  tunnelsAmbient: number,
  viewDistance: number,
  maxZoom: number,
  compass: boolean,
  autoZoom: boolean,
  swapCameras: boolean,
  toolbarPosition: ToolbarPosition,
  mapPosition: MapPosition,
  commandList: boolean,
  reverseMouse: boolean,
  panelPosition: PanelPosition,
  consoleSize: number,
  consoleDisplayTime: number,
  showHp: boolean,
  showShield: boolean,
  showElectronics: boolean,
  showAmmo: boolean,
  showExperience: boolean,
  showTemperature: boolean,
  showDynamic: boolean,
  selectionShape: SelectionShape,
  commandTooltips: boolean,
  tooltipsInResearchDialog: boolean,
  scrollbarsOnlyIfNeeded: boolean,
  descriptionType: DescriptionType,
  pauseInConstructor: boolean,
  pauseInResearchCenter: boolean,
  pauseInChangeWeapon: boolean,
  screenInfo: ScreenInfo,
  useAutosave: boolean,
  hpBarOnPanel: boolean,
  shieldBarOnPanel: boolean,
  showBuildingTooltips: boolean,
  showDamageSelection: boolean,
  scrollingMouseSpeed: number,
  scrollingKeyboardSpeed: number,
  numberOfTicksPerSecond: number,
  soundtrack: boolean,
  swapChannels: boolean,
  dontPlayProductionMessages: boolean,
  musicVolume: number,
  soundFxVolume: number,
  unitsSpeechVolume: number,
  buildingsSpeechVolume: number,
  interfaceVolume: number,
  environmentalVolume: number,
  commandFrequency: number,
  unknownNetworkParameter: number,
  mouseSensitivity: number,
  videoDoubleSized: boolean,
  videoSubtitles: boolean,
  interfaceName: string,
  autoSaveTimeMinutes: number,
  shortcuts: ShortcutData[],
}

export interface ProfileData {
  type: ProfileType,
  name: string,
  earthNetLogin: string,
  earthNetPassword: string,
  unknownCollectionValue: number,
  ucsTemplates: TemplateData[],
  unknownCollectionValue2: number,
  edTemplates: TemplateData[],
  unknownCollectionValue3: number,
  lcTemplates: TemplateData[],
  unknownCollectionValue4: number,
  enabledVideos: string[],
  gameOptions: GameOptions,
  unknownGuid: string,
  unknownString: string,
  unknownValue: number,
  unknownValue2: number,
  unknownCollectionValue5: number,
  recentIpAddresses: string[],
  unknownCollectionValue6: number,
  unknownValues: string[],
}